requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.3
decorator==5.1.1
demjson3==3.0.6
Deprecated==1.2.14
deprecation==2.1.0
diskcache==5.6.3
distro==1.8.0
docstring-parser==0.15
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
FLAML==2.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.12.1
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
joblib==1.3.2
kubernetes==28.1.0
lancedb==0.3.4
llama-index==0.9.13
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.8
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
posthog==3.1.0
prettytable==3.9.0
prompt-toolkit==3.0.36
protobuf==4.25.1
pulsar-client==3.3.0
py==1.11.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pylance==0.8.17
pymemgpt==0.2.7
pypdf==3.17.2
PyPika==0.48.9
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
questionary==2.0.1
ratelimiter==1.2.0.post0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.0
rsa==4.9
semver==3.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
